requirements.txt
streamlit==1.32.0
langchain>=0.1.9
langchain-core>=0.1.26
langchain-google-genai>=0.0.7
google-cloud-speech>=2.24.1
google-cloud-texttospeech>=2.15.1
google-generativeai>=0.3.1
python-dotenv>=1.0.1
sentence-transformers>=2.5.0
chromadb>=0.4.22
pandas>=2.1.1 